move spaces to examples and update app model
