--- components/experimental/spaces/src/index.ts
@@ -1,16 +1,0 @@
-/*!
- * Copyright (c) Microsoft Corporation. All rights reserved.
- * Licensed under the MIT License.
- */
-
-import { ContainerRuntimeFactoryWithDefaultDataStore } from "@fluidframework/aqueduct";
-import { Spaces } from "./spaces";
-
-export * from "./spaces";
-export * from "./spacesView";
-export * from "./storage";
-
-export const fluidExport = new ContainerRuntimeFactoryWithDefaultDataStore(
-    Spaces.ComponentName,
-    [[Spaces.ComponentName, Promise.resolve(Spaces.getFactory())]],
-);
