+++ examples/apps/spaces/.editorconfig
