+++ examples/apps/spaces/.eslintrc
@@ -1,11 +1,12 @@
 {
-    "extends": [
-        "@fluidframework/eslint-config-fluid/no-tslint"
-    ],
+    "extends": ["@fluidframework/eslint-config-fluid"],
     "rules": {
         "@typescript-eslint/strict-boolean-expressions": "off", // Doing undefined checks is nice
         "@typescript-eslint/unbound-method": "off", // Used to do binding for react methods
         "import/no-internal-modules": "off", // required for dynamically importing css files for react-grid-layout
-        "import/no-unassigned-import": "off" // required for dynamically importing css files for react-grid-layout
+        "import/no-unassigned-import": "off", // required for dynamically importing css files for react-grid-layout
+        "@typescript-eslint/indent": "off",
+        "no-null/no-null": "off",
+        "react/prop-types": "off"
     }
 }
