+++ examples/apps/spaces/.gitignore
