+++ examples/apps/spaces/.npmignore
