+++ examples/apps/spaces/.prettierrc
@@ -1,0 +1,5 @@
+{
+    "printWidth": 120,
+    "trailingComma": "all",
+    "arrowParens": "always"
+}
