+++ examples/apps/spaces/package.json
@@ -21,12 +21,8 @@
     "lint": "npm run eslint",
     "lint:fix": "npm run eslint:fix",
     "prepack": "npm run webpack",
-    "start": "webpack-dev-server --config webpack.config.js --package package.json",
-    "start:docker": "webpack-dev-server --config webpack.config.js --package package.json --env.mode docker",
-    "start:r11s": "webpack-dev-server --config webpack.config.js --package package.json --env.mode r11s",
-    "start:spo": "webpack-dev-server --config webpack.config.js --package package.json --env.mode spo",
-    "start:spo-df": "webpack-dev-server --config webpack.config.js --package package.json --env.mode spo-df",
-    "start:tinylicious": "webpack-dev-server --config webpack.config.js --package package.json --env.mode tinylicious",
+    "start": "webpack-dev-server",
+    "start:server": "tinylicious",
     "test": "npm run test:jest",
     "test:jest": "jest",
     "test:jest:verbose": "cross-env FLUID_TEST_VERBOSE=1 jest",
@@ -48,6 +44,7 @@
     "@fluidframework/core-interfaces": "^0.25.0",
     "@fluidframework/datastore-definitions": "^0.25.0",
     "@fluidframework/framework-interfaces": "^0.25.0",
+    "@fluidframework/get-tinylicious-container": "^0.25.0",
     "@fluidframework/map": "^0.25.0",
     "@fluidframework/protocol-definitions": "^0.1010.0",
     "@fluidframework/react-inputs": "^0.25.0",
@@ -57,6 +54,7 @@
     "@fluidframework/sequence": "^0.25.0",
     "@fluidframework/view-adapters": "^0.25.0",
     "@fluidframework/view-interfaces": "^0.25.0",
+    "@fluidframework/webpack-component-loader": "^0.25.0",
     "office-ui-fabric-react": "^7.28.1",
     "react": "^16.10.2",
     "react-collapsible": "^2.7.0",
@@ -68,7 +66,6 @@
     "@fluidframework/build-common": "^0.18.0",
     "@fluidframework/eslint-config-fluid": "^0.18.0",
     "@fluidframework/test-tools": "^0.1.0",
-    "@fluidframework/webpack-component-loader": "^0.25.0",
     "@types/node": "^10.17.24",
     "@types/react": "^16.9.15",
     "@types/react-dom": "^16.9.4",
@@ -87,6 +84,7 @@
     "eslint-plugin-react": "~7.18.0",
     "eslint-plugin-unicorn": "~15.0.1",
     "html-loader": "^0.5.5",
+    "html-webpack-plugin": "^4.3.0",
     "jest": "^24.9.0",
     "jest-junit": "^10.0.0",
     "jest-puppeteer": "^4.3.0",
