+++ examples/apps/spaces/public/index.html
@@ -1,0 +1,14 @@
+<!-- Copyright (c) Microsoft Corporation. All rights reserved. -->
+<!-- Licensed under the MIT License. -->
+
+<!DOCTYPE html>
+<html lang="en" style="height: 100%;">
+    <head>
+        <meta charset="UTF-8" />
+        <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+        <title>Test application</title>
+    </head>
+    <body style="margin: 0; height: 100%;">
+        <div id="content" style="min-height: 100%;"></div>
+    </body>
+</html>
