+++ examples/apps/spaces/src/fluid-object/spacesToolbarStyle.css
