+++ examples/apps/spaces/src/fluid-object/storage/index.ts
