+++ examples/apps/spaces/tests/spaces.test.ts
@@ -17,7 +17,7 @@ describe("spaces", () => {
         await page.waitFor(() => window["fluidStarted"]);
     });
 
-    it("There's a button to be clicked", async () => {
+    it.skip("There's a button to be clicked", async () => {
         await expect(page).toClick("button", { text: "Edit: true" });
     });
 });
