+++ examples/apps/spaces/tsconfig.json
@@ -1,0 +1,12 @@
+{
+    "extends": "@fluidframework/build-common/ts-common-config.json",
+    "exclude": ["dist", "node_modules"],
+    "compilerOptions": {
+        "target": "es6",
+        "declarationDir": "./dist",
+        "outDir": "./dist",
+        "jsx": "react",
+        "types": ["jest", "puppeteer", "jest-environment-puppeteer", "expect-puppeteer", "react", "react-dom"]
+    },
+    "include": ["src/**/*"]
+}
