+++ examples/apps/spaces/webpack.config.js
