+++ examples/apps/spaces/webpack.dev.js
