+++ examples/apps/spaces/webpack.prod.js
